Introduce providers module; refactor Gemini, Groq and Ollama
